Split date handler tests into individual files
