+++ tests/app/forms/test_fields.py
@@ -2,7 +2,6 @@ from wtforms import StringField, FormField, SelectField
 
 from app.data_model.answer_store import AnswerStore
 from app.forms.fields.max_text_area_field import MaxTextAreaField
-from app.forms.fields.month_year_date_field import MonthYearDateField
 from app.forms.fields.multiple_select_field_with_detail_answer import (
     MultipleSelectFieldWithDetailAnswer,
 )
@@ -10,11 +9,9 @@ from app.forms.fields.select_field_with_detail_answer import SelectFieldWithDeta
 from app.forms.fields.integer_field_with_separator import IntegerFieldWithSeparator
 from app.forms.fields.decimal_field_with_separator import DecimalFieldWithSeparator
 
-from app.forms.fields.date_field import DateField
 from app.forms.field_factory import get_field
 from app.forms.field_handlers.select_handler import SelectHandler
 from app.forms.error_messages import error_messages
-from app.forms.fields.year_date_field import YearDateField
 from tests.app.app_context_test_case import AppContextTestCase
 
 
@@ -78,82 +75,6 @@ class TestFields(AppContextTestCase):
         self.assertEqual(unbound_field.field_class, MaxTextAreaField)
         self.assertEqual(unbound_field.kwargs['label'], textarea_json['label'])
         self.assertEqual(unbound_field.kwargs['description'], textarea_json['guidance'])
-
-    def test_date_field(self):
-        date_json = {
-            'guidance': 'Please enter a date',
-            'id': 'period-to',
-            'label': 'Period to',
-            'mandatory': True,
-            'type': 'Date',
-            'validation': {
-                'messages': {
-                    'INVALID_DATE': 'The date entered is not valid.  Please correct your answer.',
-                    'MANDATORY': 'Please provide an answer to continue.',
-                }
-            },
-        }
-
-        with self.app_request_context('/'):
-            unbound_field = get_field(
-                date_json, error_messages, self.answer_store, self.metadata
-            )
-
-        self.assertEqual(unbound_field.field_class, DateField)
-        self.assertEqual(unbound_field.kwargs['label'], date_json['label'])
-        self.assertEqual(unbound_field.kwargs['description'], date_json['guidance'])
-
-    def test_month_year_date_field(self):
-        date_json = {
-            'guidance': '',
-            'id': 'month-year-answer',
-            'label': 'Date',
-            'mandatory': True,
-            'options': [],
-            'q_code': '11',
-            'type': 'MonthYearDate',
-            'validation': {
-                'messages': {
-                    'INVALID_DATE': 'The date entered is not valid.  Please correct your answer.',
-                    'MANDATORY': 'Please provide an answer to continue.',
-                }
-            },
-        }
-
-        with self.app_request_context('/'):
-            unbound_field = get_field(
-                date_json, error_messages, self.answer_store, self.metadata
-            )
-
-        self.assertEqual(unbound_field.field_class, MonthYearDateField)
-        self.assertEqual(unbound_field.kwargs['label'], date_json['label'])
-        self.assertEqual(unbound_field.kwargs['description'], date_json['guidance'])
-
-    def test_year_date_field(self):
-        date_json = {
-            'guidance': '',
-            'id': 'month-year-answer',
-            'label': 'Date',
-            'mandatory': True,
-            'options': [],
-            'q_code': '11',
-            'type': 'YearDate',
-            'validation': {
-                'messages': {
-                    'INVALID_DATE': 'The date entered is not valid.  Please correct your answer.',
-                    'MANDATORY': 'Please provide an answer to continue.',
-                }
-            },
-        }
-
-        with self.app_request_context('/'):
-            unbound_field = get_field(
-                date_json, error_messages, self.answer_store, self.metadata
-            )
-
-        self.assertEqual(unbound_field.field_class, YearDateField)
-        self.assertEqual(unbound_field.kwargs['label'], date_json['label'])
-        self.assertEqual(unbound_field.kwargs['description'], date_json['guidance'])
 
     def test_duration_field(self):
         date_json = {
